pyproject.toml
[project]
name = "rag-query-any-pdf"
version = "0.1.0"
description = "Query any PDF document using RAG with Inngest, FastAPI, and LlamaIndex"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.120.0",
    "inngest>=0.5.10",
    "llama-index-core>=0.14.5",
    "llama-index-readers-file>=0.5.4",
    "openai>=2.6.1",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "uvicorn>=0.38.0",
]
